refactor: route transaction controller through use cases
